+++ src/Ext/WarheadType/Body.h
@@ -39,6 +39,8 @@ public:
 		Valueable<bool> Shield_Penetrate;
 		Valueable<bool> Shield_Break;
 		Nullable<AnimTypeClass*> Shield_BreakAnim;
+		Nullable<AnimTypeClass*> Shield_HitAnim;
+		Nullable<WeaponTypeClass*> Shield_BreakWeapon;
 
 		double RandomBuffer;
 		
@@ -49,9 +51,11 @@ public:
 
 		Valueable<double> Shield_Respawn_Amount;
 		Valueable<int> Shield_Respawn_Rate;
+		Valueable<bool> Shield_Respawn_ResetTimer;
 		Valueable<int> Shield_SelfHealing_Duration;
 		Valueable<double> Shield_SelfHealing_Amount;
 		Valueable<int> Shield_SelfHealing_Rate;
+		Valueable<bool> Shield_SelfHealing_ResetTimer;
 
 		ValueableVector<ShieldTypeClass*> Shield_AttachTypes;
 		ValueableVector<ShieldTypeClass*> Shield_RemoveTypes;
@@ -88,16 +92,20 @@ public:
 			, Shield_Penetrate(false)
 			, Shield_Break(false)
 			, Shield_BreakAnim()
+			, Shield_HitAnim()
+			, Shield_BreakWeapon()
 			, Shield_AbsorbPercent()
 			, Shield_PassPercent()
 
 			, Shield_Respawn_Amount(0.0)
 			, Shield_Respawn_Rate(-1)
 			, Shield_Respawn_Rate_InMinutes(-1.0)
+			, Shield_Respawn_ResetTimer(false)
 			, Shield_SelfHealing_Duration(0)
 		    , Shield_SelfHealing_Amount(0.0)
 		    , Shield_SelfHealing_Rate(-1)
 			, Shield_SelfHealing_Rate_InMinutes(-1.0)
+			, Shield_SelfHealing_ResetTimer(false)
 			, Shield_AttachTypes()
 			, Shield_RemoveTypes()
 			, Shield_ReplaceOnly(false)